Add SET_POSTINGS to update state with postings and source
